requirements.txt
Flask
flask-session
flask-injector
jyserver
msal